update for test coverage of superbase
