+++ src/lib/supabase-client.ts
@@ -1,1 +1,1 @@
-export * from "./supabase";
+export * from "./supabase-browser";
